requirements.txt
# Styleguide
flake8==3.5.0

# Django
Django==2.0.4
django-environ==0.4.4
django-extensions==2.0.7

# Django REST Framework
djangorestframework==3.8.2

# Image handling
Pillow==5.1.0
